+++ experiments/template-instantiation/MultiplyEngineImplicit.h
@@ -8,6 +8,8 @@
 // With N translation units and 2 scalar types, that is up to 2*N compilations
 // of the same code.
 
+#include <cstdint>   // std::int32_t
+
 #include "MatrixImplicit.h"
 #include "VectorImplicit.h"
 
@@ -24,9 +26,9 @@ public:
     Vector<Val> Multiply(const Matrix<Val>& A, const Vector<Val>& x) const {
         assert(A.cols() == x.size());
         Vector<Val> y(A.rows());
-        for (std::size_t i = 0; i < A.rows(); ++i) {
+        for (std::int32_t i = 0; i < static_cast<std::int32_t>(A.rows()); ++i) {
             Val sum{0};
-            for (std::size_t j = 0; j < A.cols(); ++j) {
+            for (std::int32_t j = 0; j < static_cast<std::int32_t>(A.cols()); ++j) {
                 sum += A(i, j) * x[j];
             }
             y[i] = sum;
